Updated email address in copyright message. Removed MyDXFVec and integrated to Vec3D. Making a class for affine transformation (not done yet).

+++ src/particleiterator.hpp
@@ -23,7 +23,7 @@
  * software, please contact Berkeley Lab's Technology Transfer
  * Department at TTD@lbl.gov. Other questions, comments and bug
  * reports should be sent directly to the author via email at
- * tvkalvas@cc.jyu.fi.
+ * taneli.kalvas@jyu.fi.
  * 
  * NOTICE. This software was developed under partial funding from the
  * U.S.  Department of Energy.  As such, the U.S. Government has been
@@ -850,7 +850,7 @@ public:
 	// Cyl: x vx r vr omega
 	// 3D:  x vx y vy z vz
 	double scale_abs[PP::size()-1];
-	for( uint32_t a = 0.0; a < PP::size()-2; a+=2 ) {
+	for( uint32_t a = 0.0; a < (uint32_t)PP::size()-2; a+=2 ) {
 	    scale_abs[a+0] = 1.0;
 	    scale_abs[a+1] = 1.0e6;
 	}
@@ -1061,6 +1061,9 @@ public:
 
 
 #endif
+
+
+
